Fix test helpers for in memory server
In order to use a ghttp server as a testing fake in our integration
tests, we need to be able to make paths and queries to particular return
values, rather than just paths. This adds a test helper that registers
stateful handlers with the underlying ghttp server so that we can
achieve these more detailed fakes.

[#163539592]